# swagger/dist/build.yaml
openapi: 3.0.0
info:
  description: DXDATA의 스마트 관제시스템(Buoy) API입니다.
  version: 1.0.0
  title: Smart Control System API
  contact:
    email: dxdata@dxdata.co.kr
    url: 'http://www.dxdata.co.kr'
basePath: /
...
paths:
  /main/data:
    get:
      tags:
        - main
      description: |
        # 현재 자신의 위치기반으로 데이터를 가져옵니다.

        ## 기상청에서 가져오는 기상 정보입니다.

            "meteo_val": {

            "data": {
            
              "humidity": 습도 %
              
              "rain_amt_hour": 강수량 없으면 강수없음, 그외에는 mm

              "rain_code": 강수코드, 0 : 없음, 1 : 비, 2 : 비/눈, 3 : 눈, 5 : 빗방울, 6 : 빗방울 눈랄림, 7: 눈날림
            
              "sky": 하늘상태, 1 : 맑음, 3 : 구름많음, 4 : 흐림

              "sn_wind": 남북 바람 m/s

              "temperature": 온도

              "thunder": 낙뢰 코드,
            
              "we_wind": 동서 바람 m/s,

              "wind_direction": 풍향 deg,

              "wind_velocity": 퐁숙 m/s

            },

            "region" : 지역


        ## 관측부이에서 가져오는 해상정보들입니다.


            "obs_data": {

              "Salinity": 염분 psu

              "air_press": 기압 hpa

              "air_temp": 기온

              "record_time": 측정 시간

              "tide_level": 조위 cm

              "water_temp": 수온

              "wind_dir": 풍향 deg

              "wind_gust": 돌풍 m/s

              "wind_speed": 풍속 m/s

            }


        ## 조류 속도입니다.

            "tidal": {

              "current_direct": 방향 deg,

              "current_speed": 속도 cm/s

            }


        ## 파고입니다.

            "wave_hight": {

              "record_time": 측정시간

              "wave_height": 파고 m

            }
      parameters:
        - in: query
          name: latitude
          schema:
            type: string
          example: 35.1513466
          required: true
        - in: query
          name: longitude
          schema:
            type: string
          example: 128.1001125
          required: true
      responses:
        '200':
          description: 메인 지역별 현재 데이터
          content:
            apllication/json:
              schema:
                $ref: '#/components/schemas/MainData'
      security:
        - JWT: []
  /main/group:
    get:
      tags:
        - main
      description: 메인에서 각 부이 그룹별 현재 데이터를 가져옵니다.
      responses:
        '200':
          description: 각 부이 그룹별 현재 데이터
          content:
            apllication/json:
              schema:
                $ref: '#/components/schemas/BuoyGroupMain'
      security:
        - JWT: []
  /main/group/total:
    get:
      tags:
        - main
      description: 메인에서 부이 그룹의 전체 평균 데이터를 가져옵니다.
      responses:
        '200':
          description: 각 부이 그룹의 전체 평균 데이터
      security:
        - JWT: []
  /main/warn:
    get:
      tags:
        - main
      description: 메인에서 각 그룹별 경고 리스트를 가져옵니다.
      responses:
        '200':
          description: 각 부이 그룹별 현재 경고 리스트
          content:
            apllication/json:
              schema:
                $ref: '#/components/schemas/WarnList'
      security:
        - JWT: []
...
  /main/region:
    get:
      tags:
        - legacy
      description: '메인에서 지역별 현재 데이터를 가져옵니다. 파고는 m. 조류 속도는 cm/s. 풍속 돌풍은 m/s, 기압은 hPa, 조위는 cm, 염분은 psu, 그외의 방향은 deg입니다.'
      parameters:
        - in: query
          name: location
          schema:
            type: string
            enum:
              - tongyeong
              - geojedo
          required: true
      responses:
        '200':
          description: 메인 지역별 현재 데이터
          content:
            apllication/json:
              schema:
                $ref: '#/components/schemas/MainData'
...
components:
  schemas:
    MainData:
      type: object
      properties:
        meteo_val:
          type: object
        obs_data:
          type: object
        tidal:
          type: object
        wave_hight:
          type: object
      example:
        meteo_val:
          data:
            humidity: 15
            rain_amt_hour: 강수없음
            rain_code: 0
            sky: 1
            sn_wind: 2
            temperature: 24
            thunder: 0
            we_wind: 0.1
            wind_direction: 205
            wind_velocity: 2
          region: 진주시
        obs_data:
          Salinity: '33.7'
          air_press: '1013.7'
          air_temp: '14.5'
          record_time: '2022-04-13 15:13:00'
          tide_level: '122.0'
          water_temp: '13.9'
          wind_dir: '112.0'
          wind_gust: '6.6'
          wind_speed: '3.4'
        tidal:
          current_direct: '215.8'
          current_speed: '34.0'
        wave_hight:
          record_time: '2022-04-13 15:00'
          wave_height: '2.3'
    BuoyGroupMain:
      type: object
      required:
        - group_id
        - group_name
        - group_latitude
        - group_longitude
        - group_water_temp
        - group_salinity
        - group_height
        - group_weight
        - plain_buoy
        - smart_buoy
        - region
      properties:
        group_id:
          type: float
        group_name:
          type: string
        group_latitude:
          type: double
        group_longitude:
          type: double
        group_water_temp:
          type: float
        group_salinity:
          type: float
        group_height:
          type: float
        group_weight:
          type: float
        plain_buoy:
          type: integer
        smart_buoy:
          type: integer
        region:
          type: string
      example:
        group_id: 1
        group_name: A
        group_latitude: 32.3333
        group_longitude: 32.4444
        group_water_temp: 3.22
        group_salinity: 3.5
        group_height: 5.4
        group_weight: 2.3
        plain_buoy: 100
        smart_buoy: 30
        region: 통영시
    WarnList:
      type: object
      required:
        - group_id
        - group_name
        - line
        - warn_type
        - message
      properties:
        group_id:
          type: integer
        group_name:
          type: string
        line:
          type: integer
        warn_type:
          type: string
        message:
          type: string
...
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT